Przed poprawieniem scenemanager

+++ src/main/java/com/reservio/reservation_system/reservationsystemdesktop/model/room/RoomDto.java
@@ -1,10 +1,12 @@
 package com.reservio.reservation_system.reservationsystemdesktop.model.room;
 
+import java.math.BigDecimal;
+
 public record RoomDto(
         Long id,
         Long number,
         String name,
         Integer capacity,
-        float pricePerNight,
+        BigDecimal pricePerNight,
         String status
 ) {}
